postgres_to_es/admin_panel/requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-split-settings==1.1.0
elastic-transport==8.1.2
elasticsearch==8.2.0
flake8==3.8.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9
py==1.11.0
pycodestyle==2.6.0
pydantic==1.9.1
pyflakes==2.2.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uWSGI==2.0.20
